Organize and minor clean up code
