+++ test/TextRenderingTest/TypographyTest.cs
@@ -1,10 +1,6 @@
 ﻿using ImGui;
 using System;
-using System.Diagnostics;
 using Xunit;
-using System.Collections.Generic;
-using System.Text;
-using Typography.Rendering;
 using System.IO;
 using Typography.OpenFont;
 
@@ -29,6 +25,7 @@ namespace TextRenderingTest
         {
             var glyph = typeFace.GetGlyphByIndex(36);
             var points = glyph.GlyphPoints;
+            //TODO check points
         }
     }
 }
